refactor(core): move usage-windows into usage/; de-triplicate windowKey
The window-key format lived in three places — telemetry/usage-windows plus a
private windowKey() copied into the Prisma and DocumentDB repos (each with a
"mirrors telemetry/usage-windows" apology comment). Move usage-windows to usage/
(its true layer — the application-layer usage-commands was reaching across into
telemetry/ for it), export a shared windowKey, express computeUsageWindows on top
of it, and delete both private copies. One source of truth for the write/read
window format.
